--- client/python/armada_client/.python-version
@@ -1,3 +1,0 @@
-3.7.13
-3.8.13
-3.9.12
